api/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx

# Phase 4 — real multimodal ingestion
Pillow
faster-whisper
sentence-transformers
open_clip_torch
transformers
